requirements.txt
torch==2.4.0
transformers>=4.40.0
accelerate>=0.30.0
pillow>=10.0.0
huggingface_hub>=0.23.0
protobuf>=4.25.0
numpy<2.0
runpod>=1.6.0
flash-attn>=2.5.0